本登録フロントエンド完成

+++ routes/api/logins/register.js
@@ -1,38 +1,55 @@
-import { isMail, isUserDuplication, isMailDuplication, isMailDuplicationTemp, isCsrftoken, sendMail} from "../../../util/takoFunction.ts";
+import { isMail, isMailDuplication, isMailDuplicationTemp, sendMail} from "../../../util/takoFunction.ts";
 import tempUsers from "../../../models/tempUsers.js";
+import { load } from "https://deno.land/std@0.204.0/dotenv/mod.ts";
+const env = load();
+const secretKey = "6LdiCBopAAAAAOmECv_L9EYRQyuFBRiRLEKSN3ES"
 export const handler = {
-  async POST(req,res, ctx) {
+  async POST(req) {
       const data = await req.json();
       const email = await data.mail;
       const ismail = isMail(email)
+      const rechapcha = await data.token
+      const isSecsusRechapcha = await fetch(`https://www.google.com/recaptcha/api/siteverify?secret=${secretKey}&response=${rechapcha}`)
+      const score = await isSecsusRechapcha.json()
+      if(score.score < 0.7) {
+        return new Response(JSON.stringify({"status": "rechapchaerror"}), {
+          headers: { "Content-Type": "application/json",
+                      status : 403},
+        });
+      }
       const ismailduplication = await isMailDuplication(email)
       if(ismail) {
         if(!ismailduplication) {
             try {
             const key = generateRandomString(255);
             if(isMailDuplicationTemp(email)) {
-              tempUsers.deleteOne({mail: email})
+              await tempUsers.deleteOne({mail: email})
             }
             await tempUsers.create({mail: email, key: key})
             sendMail(email,"本登録を完了してください",`https://takos.jp/register?key=${key}`)
             return new Response(JSON.stringify({status: true}), {
-              headers: { "Content-Type": "application/json" },
+              headers: { "Content-Type": "application/json",
+                          status : 200 },
             });
             } catch (error) {
-              return new Response(JSON.stringify({"status": "error"}), {
-                headers: { "Content-Type": "application/json" },
+              return new Response(JSON.stringify({"status": "500error"}), {
+                headers: { "Content-Type": "application/json",
+                            status : 403},
               });
             }
         }else {
-          return new Response(JSON.stringify({"status": "error"}), {
-            headers: { "Content-Type": "application/json" },
+          return new Response(JSON.stringify({"status": "mailerror"}), {
+            headers: { "Content-Type": "application/json",
+                        status : 403},
           });
         }
       }else {
-        res.status(403).send({"status": "error"})
-        /*return new Response(JSON.stringify({"status": "error"}), {
-          headers: { "Content-Type": "application/json" },
-        });*/
+        headers.set("Content-Type", "application/json");
+        return new Response(JSON.stringify({"status": "error"}), {
+          headers: { "Content-Type": "application/json",
+                      status : 403
+                    },
+        });
       }
   }
 };
